Separate tweet param & user param mapper utilities
